+++ src/lazily/__init__.py
@@ -58,7 +58,7 @@ __all__ = [
     "slot",
     "slot_def",
 ]
-__version__ = "0.12.0"
+__version__ = "0.13.0"
 
 from . import ipc
 from .cell import Cell, CellSlot, cell, cell_def